move wizard to subdir per @elecpower 's comment in #6556
