+++ companion/src/wizard/wizarddialog.cpp
