+++ companion/src/wizard/wizarddialog.h
